requirements.txt
# Google Agent Development Kit
google-adk>=1.0.0

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Environment variables
python-dotenv>=1.0.0

# PDF Report Generation
markdown-pdf>=1.3.0

# Database
# SQLite is built-in, no additional package needed

# Optional: For enhanced observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
